login/register request dto
