--- src/main/java/com/chokchok/chokchokapi/member/dto/MemberLoginRequestDto.java
@@ -1,7 +1,0 @@
-package com.chokchok.chokchokapi.member.dto;
-
-public record MemberLoginRequestDto(
-        String email,
-        String password
-) {
-}
